Move TLSClient to http package
