+++ receivers/util_test.go
@@ -1,8 +1,6 @@
 package receivers
 
 import (
-	"crypto/tls"
-	"net/http"
 	"testing"
 
 	"github.com/stretchr/testify/require"
@@ -105,31 +103,6 @@ func TestNewTLSConfig(t *testing.T) {
 					require.NotEmpty(t, tlsCfg.Certificates, "expected Certificates to be set, but it was empty")
 				}
 			}
-		})
-	}
-}
-
-func Test_NewTLSClient(t *testing.T) {
-	tc := []struct {
-		name   string
-		cfg    *tls.Config
-		expCfg *tls.Config
-	}{
-		{
-			name:   "empty TLSConfig",
-			expCfg: &tls.Config{Renegotiation: tls.RenegotiateFreelyAsClient},
-		},
-		{
-			name:   "valid TLSConfig",
-			cfg:    &tls.Config{InsecureSkipVerify: true},
-			expCfg: &tls.Config{InsecureSkipVerify: true},
-		},
-	}
-
-	for _, tt := range tc {
-		t.Run(tt.name, func(t *testing.T) {
-			c := NewTLSClient(tt.cfg)
-			require.Equal(t, tt.expCfg, c.Transport.(*http.Transport).TLSClientConfig)
 		})
 	}
 }
